Split up CallAll into common + shim
